Add fixed_frame_rate_flag constraint to AVC VUI
We do not support the carriage of variable frame rate content, and
without this flag set properly in the SPS its potentially impossible for
a decoder to know the rate at which to decode the bitstream. Of course,
such information is also carried at the MXF layer, but we want to ensure
that it is available to the AVC decoder.